Attempt to upgrade to react-scripts 2.0
Also made the package.json a bit more sane.
I have not been able to test it, since it won't build.
Not sure if this is because I am running windows or if the build is
wrong.

I am giving up on trying to make it build, you can give it a try and do some regression testing,
or throw it away if it is not working.